Revert "Merge branch 'BALS' into refactor-Fileforest-to-Filetree"
This reverts commit dd9e72748406de76e2fee83d109664bfd9fd9943, reversing
changes made to 4e7d38b68c74c47ec6c362872fb6b38ce4fd691d.

+++ src/components/Filetree.vue
@@ -1,18 +1,18 @@
 <template>
 	<div>
-		<ul v-for="item in filetree" :key="item.id">
+		<ul v-for="item in filetree.objects" :key="item.id">
 			<li
 				:class="{
 					'list-item': true,
 					'is-folder': item.type === 'folder',
 				}"
+				@dragenter.prevent="handleDragover"
 				@dragover.prevent="handleDragover"
 				@dragleave.prevent="handleDragleave"
 				@drop.prevent="handleDropEvent($event, path, item)"
 				@click="toggleExpansion(item)"
 			>
 				<FiletreeEntry
-					:id="item.id"
 					:icon="
 						item.type === 'file'
 							? 'insert_drive_file'
@@ -20,9 +20,7 @@
 							? 'folder_open'
 							: 'folder'
 					"
-					:name="item.name"
-					:state="item.state"
-					:progress="item.progress"
+					:file="item"
 					:read-only="parentState === 'deleted'"
 					@delete="deleteEntry"
 					@restore="restoreEntry"
@@ -31,7 +29,7 @@
 			<li v-if="item.type === 'folder'">
 				<Filetree
 					v-show="expandedFolders.includes(item.id)"
-					:filetree.sync="item.objects"
+					:filetree.sync="item"
 					:value="value"
 					:path="path + '/' + item.name"
 					:parent-state="item.state"
@@ -55,7 +53,7 @@ export default {
 	mixins: [upload, filetree],
 	props: {
 		filetree: {
-			type: Array,
+			type: Object,
 			required: true,
 		},
 		value: {
@@ -82,14 +80,41 @@ export default {
 				return;
 			}
 			if (to === "deleted") {
-				this.filetree.forEach((item) => {
-					this.deleteEntry(item.id, item.name);
+				this.filetree.objects.forEach((item) => {
+					if (item.state !== "deleted") {
+						this.deleteEntry(item.id, item.name);
+					}
 				});
 			} else if (!to) {
-				this.filetree.forEach((item) => {
-					this.restoreEntry(item.id, item.name);
+				this.filetree.objects.forEach((item) => {
+					if (item.state === "deleted") {
+						this.restoreEntry(item.id, item.name);
+					}
 				});
 			}
+		},
+		filetree: {
+			handler: function(to, from) {
+				if (to.type === "folder") {
+					const folderProgress = to.objects.reduce((progress, item) => {
+						return item.progress
+							? progress
+								? (progress + item.progress) / 2
+								: item.progress
+							: progress;
+					}, 0);
+					this.filetree.progress = folderProgress ? folderProgress : undefined;
+					if (to.state !== "deleted") {
+						if (to.objects.some((file) => !!file.state)) {
+							this.filetree.state = "modified";
+						} else {
+							this.filetree.state = undefined;
+						}
+					}
+					this.$forceUpdate();
+				}
+			},
+			deep: true,
 		},
 	},
 	methods: {
@@ -103,20 +128,40 @@ export default {
 				this.expandedFolders.push(id);
 			}
 		},
+		getItemIndex(name) {
+			return this.filetree.objects.findIndex((item) => item.name === name);
+		},
+		abortRequest(file) {
+			if (file.xhr) {
+				file.xhr.abort();
+			}
+			return file;
+		},
 		deleteEntry(id, name) {
-			const itemIndex = this.filetree.findIndex((item) => item.name === name);
-			const item = this.filetree[itemIndex];
-			// already in list
-			if (item.state === "deleted" || item.state === "updated") {
-				return; // shouldn't be possible
-				//return console.warn("deleting this item shouldn't be possible");
+			const itemIndex = this.getItemIndex(name);
+			this.abortRequest(this.filetree.objects[itemIndex]);
+			const item = this.filetree.objects[itemIndex];
+			if (item.state === "modified") {
+				const hasModifiedChilds = item.objects.some((file) =>
+					["new", "updated", "modified"].includes(file.state)
+				);
+				if (hasModifiedChilds && !confirm(this.$lang.upload.confirmDelete)) {
+					return;
+				}
+			}
+			if (item.state === "deleted") {
+				console.warn("deleting this item shouldn't be possible");
+			} else if (item.state === "updated") {
+				this.restoreEntry(id, name);
+				this.deleteEntry(id, name);
 			} else if (item.state === "new") {
+				// don't save anymore
 				this.value.save.splice(this.value.save.indexOf(id), 1);
-				this.filetree.splice(itemIndex, 1);
+				this.filetree.objects.splice(itemIndex, 1);
 			} else {
 				// save files
-				this.filetree[itemIndex].state = "deleted";
-				if (this.filetree[itemIndex].type === "file") {
+				this.filetree.objects[itemIndex].state = "deleted";
+				if (this.filetree.objects[itemIndex].type === "file") {
 					this.value.delete.push(id);
 				}
 			}
@@ -124,65 +169,77 @@ export default {
 			this.$forceUpdate();
 		},
 		restoreEntry(id, name) {
-			const itemIndex = this.filetree.findIndex((item) => item.name === name);
-			const item = this.filetree[itemIndex];
+			const itemIndex = this.getItemIndex(name);
+			this.abortRequest(this.filetree.objects[itemIndex]);
+			const item = this.filetree.objects[itemIndex];
 			if (item.state === "new") {
+				console.error("unhandled state");
 				return; // shouldn't be possible
 			}
 			if (item.state === "updated") {
-				this.filetree[itemIndex].state = undefined;
+				this.filetree.objects[itemIndex].id = item.originalId;
+				this.filetree.objects[itemIndex].originalId = undefined;
+				this.filetree.objects[itemIndex].state = undefined;
 
 				this.value.save
-					.filter((savedId) => {
-						let res = savedId.includes(id);
-						if (res) return true;
-					})
+					.filter((savedId) => savedId.includes(id))
 					.forEach((elem) => {
 						this.value.save.splice(this.value.save.indexOf(elem), 1);
 					});
 			}
 			if (item.state === "deleted") {
-				this.filetree[itemIndex].state = undefined;
+				this.filetree.objects[itemIndex].state = undefined;
 				this.value.delete.splice(this.value.delete.indexOf(id), 1);
 			}
 			this.$emit("update", this.value);
 			this.$forceUpdate();
 		},
 		handleDropEvent(event, prefix, item) {
-			if (item.type === "folder") {
-				this.dropFile(event, prefix + "/" + item.name).then(
+			if (item.type === "folder" && item.state !== "deleted") {
+				this.$_dropFile(event, prefix + "/" + item.name).then(
 					(newItemsForest) => {
-						const srcTree = this.filetree;
-						this.recursiveSaveAfterUpload(newItemsForest);
+						const srcTree = this.filetree.objects;
+						this.$_recursiveSaveAfterUpload(newItemsForest);
 						const currentItemIndex = srcTree.findIndex(
 							(node) => node.name === item.name
 						);
 						if (currentItemIndex === -1) {
-							srcTree[currentItemIndex].objects.push(
-								this.recursiveSetState(newItemsForest, "new")
+							this.filetree.objects[currentItemIndex].objects.push(
+								this.$_recursiveSetState(newItemsForest, "new")
 							);
 						} else {
-							srcTree[currentItemIndex].objects = this.mergeIntoTree(
-								srcTree[currentItemIndex].objects,
+							this.filetree.objects[
+								currentItemIndex
+							].objects = this.$_mergeIntoForest(
+								this.filetree.objects[currentItemIndex].objects,
 								newItemsForest
 							);
 						}
 						this.$emit("update", this.value);
-						this.$emit("update:filetree", srcTree);
+						this.$emit("update:filetree", this.filetree);
 					}
 				);
 			}
 			this.handleDragleave(event);
 		},
 		wrappingFolder(event, callback) {
+			// TODO refactor, don't use direct dom manipulation
+			// TODO highlight dropzone only if folder isn't deleted
 			const wrapper = event.target.closest(".list-item");
 			if (Array.from(wrapper.classList).includes("is-folder")) {
 				callback(wrapper);
+			} else {
+				callback(undefined);
 			}
 		},
 		handleDragover(event) {
 			this.wrappingFolder(event, (wrapper) => {
-				wrapper.classList.add("dragover");
+				if (wrapper && !wrapper.querySelector(".is-deleted")) {
+					wrapper.classList.add("dragover");
+					event.dataTransfer.dropEffect = "copy";
+				} else {
+					event.dataTransfer.dropEffect = "none";
+				}
 			});
 		},
 		handleDragleave(event) {
@@ -200,9 +257,10 @@ ul {
 	margin: 0;
 	li {
 		list-style-type: none;
+		border-radius: 2px;
 		transition: background-color 0.1s linear 0.1s;
 		&.dragover {
-			background-color: rgb(154, 243, 191);
+			background-color: #ccc;
 			transition: background-color 0.1s linear;
 		}
 	}
